+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
 Changelog for Mantarray Frontend Components
 ===========================================
+
+0.3.0 (unreleased)
+------------------
+
+- Added y-axis-zoom for waveform display
+- Added heatmap
+- Added websocket to receive waveform data asynchronously and in real time
+
 0.2.0 (2021-04-01)
 ------------------